payment-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
pydantic==2.4.2
grpcio==1.59.3
grpcio-tools==1.59.3
requests==2.31.0
httpx==0.25.2
prometheus-client==0.21.1
